refactor(ci): move backend jobs into dedicated folder
