+++ .circleci/ci/src/jobs/index.ts
@@ -1,13 +1,10 @@
 export * from './job-add-docker-images-in-snyk';
-export * from './job-build-backend';
+export * from './backend/job-build-backend';
 export * from './job-build-backend-images';
-export * from './job-community-build-backend';
 export * from './job-danger-js';
 export * from './job-deploy-on-azure';
-export * from './job-nexus-staging';
 export * from './job-package-bundle';
 export * from './job-perf-lint-build';
-export * from './job-publish';
 export * from './job-publish-pr-env-urls';
 export * from './job-publish-prod-docker-images';
 export * from './job-publish-rpm-packages';
@@ -15,4 +12,3 @@ export * from './job-release-notes-apim';
 export * from './job-setup';
 export * from './job-slack-announcement';
 export * from './job-sonarcloud-analysis';
-export * from './job-validate';
